Added an intersectCounter to kdtree to see how many intersections are performed.
The problem seems to be that, in case of the "cubeStraight.obj" all 12 primitives are checked for one traversal. Atleast half of the primitives or nodes should be skipped.

Debugging this error should fix the long execution times for the kdtree.

The program was run for bunnBlender_v2.obj and it took 2237 seconds and provided a bunny output. So the actual intersections and camera system 'seem' to be fine.