Created a parse objects by manually filling the first object in the cubeStraight.obj file.
The output produced is different. There is some error in the way the objects are being read.

The current output (for both methods) shows an output simillar to Blender. However in Blender the entire triangle is in view, whereas in this ray-tracer the bottom is cut-off. Must be some error in perspective that needs to be fixed.