Added a Sphere with radius 100 at Z -100.5 to make it look like a plane.
Compared to book the "plane" Sphere is on the top, which means there the flipping is confirmed and has to be corrected before going further.

In the TraceRay_debug() it is necessary to dynamic_cast the ObjectClass array to SphereClass to access the GetCenter() function. This must be changed if the function is kept for the future. Although this is only for debugging purposes so maybe it is fine.