The invalid argument error still exists. Unfixed. Tried setting a different file name. Same error. Happens every other time.
In TraceRay_debug() there are two variables t_debug and currentLowestVal_debug. t_debug is to store the intersection value returned by GetIntersection()

currentLowestVal_debug is SUPPOSED to store the lowest intersection value (i.e. t_debug) returned among ALL objects.

In the code from line 413 to the end of the function, the value of t_debug value was being used ALWAYS. This is incorrect since that would only be correct if the LAST intersection value of t_debug was the right one.

However if there is intersection of earlier objects which were lower, then that lower value would be stored in currentLowestVal_debug.

After this t_debug was replaced with currentLowestVal_debug AFTER the intersection for loops.

However always using the value of currentLowestVal_debug is wrong when there are no intersections with ANY object.

This same issue might exist in the other TraceRay functions. Where 't_debug' is denoted by 'omega' and 'currentLowestVal_debug' is denoted by 'currentLowestVal'. The lowest value of the intersection is not properly used.

IMMEDIATE FIX NEEDED.