Working on TraceRay(), the brute force method.
First there is an iteration over every object with the current ray. The result of the intersection is stored in 'omega'. The lowest intersection value is stored in 'currentLowestVal'

After this for loop iteration, in the next section we check for shadows. In line 240 we check to make sure that the intersection value is not -1.

Now in line 246 there was no check earlier, but now we test to make sure that the currentLowestVal is not 'infinity' and greater than 'epsilonval_small'. The infinity check may be redundant since it is covered by checking if the closest = -1, i.e. no intersections were found.

In line 271 the check for shadowOmega was changed to 'epsilonval_small' instead of a magic number.
- Here we probably do not need to do the 'infinity' check since shadowOmega is initialized to 0.0 and the shadows are additive. So the exact lowest value is not needed.