6ca29968d457c56872add18898f1b4867fe77b04 (Initializing the counter for screenHeight, i.e. the variable 'j', from 0 flips the image back., 2021-10-14)
mentioned that the resultant image seemed to be flipped as compared to the one in the rt1weekend book. So the pixel counter variable for imageHeight (or screenHeight for earlier commits), "j", was incremented, instead of being decremented like in the book.

However the image actually was being flipped, and reverting to exactly how it's in the book, i.e. j starts at imageHeight and is decremented, gives the correct output.

In addition, the sphere with the radius of 100 being used blocks all other spheres. Need to debug.