Finished the implementation of the ppm file saver for ColourClass.
The outpt is now not flipped. So the suscpicion that the bmp saver was somehow flipping the image was correct.

Also added a timer to count the file write time.

Writing to a ppm file is far slower than the bmp. Takes almost 20 seconds to write the ppm file.

Debug and check the bmp saver later for why the image gets flipped, since it is faster.