The null reflectKr issue was solved by adding a Illumination class to Sphere.
Another error showed up with a null pointer error for a Colour value.
This happened because during the first recursion for a reflected ray in the depth < maxDepth if condition, when TraceRay_debug() was called again, the currentLowestVal_debug during that recursion was not equal to infinity but ALSO NOT greater than epsilonval_small. And when the function finished, nothing was being returned.

This is something that can be caught and solved by unit testing, since a function with a return value should always return something, never null.

However maybe the 'currentLowestVal > epsilonval_small' check is not correct; since the output is weird. The areas that should be bright are dark and vice vera.