Debugging tip SO- Adrian McCarthy
In your case, the sphere itself doesn't exhibit the acne, but its reflection does. That suggests that the nudge is sometimes smaller than the cumulative loss of precision from the initial ray to the reflection ray to the sphere. Maybe your nudge is too small. Maybe something goes wrong where you compute the reflection ray.

It could also be that the intersection with the plane is sometimes on the wrong side, so your reflection ray heads off below the floor instead of back up.

To debug, I'd hack the "I hit nothing" color to hot pink and the "I'm in the shadow" color to lime green. If you see pink in those spots, then something went wrong computing the reflection ray. If you see lime green, then it's surface acne and you need to ensure you don't have a systematic precision loss and that your nudge factor is sufficient.