Matched the output of the red sphere in rt1w.
The book basically just draws a red pixel if it intersects with an imaginary sphere placed at z = -1.

While no actual sphere is "created", just before the lerp Colour function there is a test to check if the ray will intersect with a sphere at z = -1. If yes it just draws a red pixel and skips the lerp function.