Changed the code so the output only had blue for background and red for the sphere. The output still did not change.
So the issue was probably not with the intersection or the normals at all.

Upon further inspection the issue was not converting the RGB value in the pixelColour to the 0-255 range. This led to the RGB values to be extremely low as they are in the 0-1 range by default.

Changing that the current output is of blue background with red sphere as expected.