The issue was the declaration of the 'cam' CmaeraClass variable in main.
To check with the earlier basic Sphere output it was declared without the parameterized constructor. However the lowerLeftCorner, horizontal, and vertical calculations were only done in the newly created parameterized constructor.

As that constructor was not being called, the direction was never set.

To test for basic output calling that cam variable with dummy 0 values calls the parameterized constructor which should solve the issue.