Changed the setting of the direction variable in all the RayClass constructor to use the setRayDirection() function.
Found an interesting possible bug. When calculating the inverse of the direction there is no check for if the direction component is zero. 1 / 0 is undefined/inf.

Where exactly is inverse of vector used and what exactly does the inverse of a vector mean?

Currently the code to auto set the inverseDirection is commented out.