The reason the spheres are all black because the maxTerm is 0. So the negative value gets clamped to 0. This exists to prevent thenegative dot product of the light and normal for when the light is behind the surface, i.e. the normal and light are on opposite sides of the intersecting surface.
However to calculate the currLightDirection there is a multiplication of the direction with -1 to make it the same direction as the normal.

Removing this negative 1 makes the output a 'little' lighter and the three spheres are distinguishable in the output albeit barely.

To counter this we need to look  at both methods in scratchapixel and rt1w. In rt1w there is a method to always store the normal to the sphere in a particular way which is the current GetIntersction() code is commented out.

First resolve that. It may solve the current issue. However what if application of a diffuse surface leads to a black output for most surface materials? Phong model uses diffuse as one of it's components so that maybe necessary to display a correct output?